Move GenerateKeyAndAddr to helper package since it's useful for other tests, not only e2e ones
