+++ e2e/encoding_test.go
@@ -6,11 +6,12 @@ import (
 	"time"
 
 	"github.com/0xPolygon/minimal/e2e/framework"
+	"github.com/0xPolygon/minimal/helper/tests"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestEncoding(t *testing.T) {
-	_, from := framework.GenerateKeyAndAddr(t)
+	_, from := tests.GenerateKeyAndAddr(t)
 
 	srvs := framework.NewTestServers(t, 1, func(config *framework.TestServerConfig) {
 		config.SetConsensus(framework.ConsensusDev)
